py-server/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
web3>=6.0.0
torch>=2.0.0
numpy>=1.21.0
ezkl>=7.0.0
onnx>=1.12.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
websockets>=10.0
sentence-transformers>=2.2.2
scikit-learn>=1.0.0 